Docs(fe): API 기능별 파일 분리
